% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.fh.R
\name{compare}
\alias{compare}
\title{Shows plots for the comparison of estimates}
\usage{
compare(object, indicator, MSE, CV, ...)
}
\arguments{
\item{indicator}{optional character vector that selects which indicators
shall be returned: (i) all calculated indicators ("all");
(ii) each indicator name: "Mean", "Quantile_10", "Quantile_25", "Median",
"Quantile_75", "Quantile_90", "Head_Count",
"Poverty_Gap", "Gini", "Quintile_Share" or the function name/s of
"custom_indicator/s"; (iii) groups of indicators: "Quantiles", "Poverty",
"Inequality" or "Custom".If two of these groups are selected, only the first
one is returned. Defaults to "all". Note, additional custom indicators can be
defined as argument for model-based approaches (see also \code{\link{ebp}})
and do not appear in groups of indicators even though these might belong to
one of the groups.}

\item{direct}{optional, an object of type "emdi","direct", representing point
and MSE estimates.}

\item{model}{an object of type "emdi","model", representing point and MSE
estimates.}

\item{label}{argument that enables to customize title and axis labels. There
are three options to label the evaluation plots: (i) original labels ("orig"),
(ii) axis lables but no title ("no_title"), (iii) neither axis
labels nor title ("blank").}

\item{color}{a vector with two elements. The first color determines
the color of the line in the scatter plot and the color for the direct
estimates in the line plot. The second color specifies the color of the line
for the model-based estimates.}

\item{shape}{a numeric vector with two elements. The first shape determines
the shape of the points in the line plot for the direct estimates and the
second shape for the model-based estimates. The options are numbered from
0 to 25.}

\item{line_type}{a character vector with two elements. The first line type
determines the type of the line for the direct estimates and the
second type for the model-based estimates. The options are: "twodash",
"solid", "longdash", "dotted", "dotdash", "dashed" and "blank".}

\item{gg_theme}{\code{\link[ggplot2]{theme}} list from package \pkg{ggplot2}.
For using this argument, package \pkg{ggplot2} must be loaded via
\code{library(ggplot2)}. See also Example 2.}
}
\value{
A scatter plot and a line plot comparing direct and model-based
estimators for each selected indicator obtained by \code{\link[ggplot2]{ggplot}}.
}
\description{
For all indicators or a selection of indicators two plots are returned. The
first plot is a scatter plot of estimates to compare and the second is a line
plot with these estimates.
}
