% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_plots.R
\name{CV_plots}
\alias{CV_plots}
\title{CV plots}
\usage{
CV_plots(fit_FH, label_direct = "Direct", label_FH = "FH-MI",
  line = 20, colline = "red", color = c("dodgerblue4", "#99CC00"))
}
\arguments{
\item{fit_FH}{fitted FH model.}

\item{label_direct}{label for the direct estimates.}

\item{label_FH}{label for FH estimates.}

\item{line}{benchmark CV level.}

\item{colline}{color of benchmark line.}

\item{color}{color of the two estimates.}
}
\value{
two CV plots.
}
\description{
This function returns two different plots for the CV.
}
