% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH.R
\name{fh}
\alias{fh}
\title{Function for Fay-Herriot model}
\usage{
fh(fixed, vardir, combined_data, domains = NULL, method = "reml",
  interval = c(0, 1000), transformation = "no", eff_smpsize = NULL,
  MSE = "analytical", B = NULL, alpha = 0.05)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators.}

\item{vardir}{a character string indicating the name of the variable containing
the domain-specific sampling variances of the direct estimators that are
included in \cr \code{combined_data}.}

\item{combined_data}{a data set containing the direct estimates,
the sampling variances, the explanatory variables and the domains.}

\item{domains}{a character string indicating the domain variable that is
included in \cr \code{combined_data}. If \code{NULL}, the domains are numbered
consecutively.}

\item{method}{a character string describing the method for the estimation of
the variance of the random effects. Methods that can be chosen
(i) restricted maximum likelihood (REML) method ("\code{reml}"),
(ii) maximum likelihood method ("\code{ml}"),
(iii) adjusted REML following \cite{Li and Lahiri (2010)} ("\code{amrl}"),
(iv) adjusted ML following \cite{Li and Lahiri (2010)} ("\code{ampl}"),
(v) adjusted REML following \cite{Yoshimori and Lahiri (2014)}
("\code{amrl_yl}"),
(vi) adjusted ML following \cite{Yoshimori and Lahiri (2014)}
("\code{ampl_yl}").
 Defaults to "\code{reml}".}

\item{interval}{interval for the estimation of sigmau2.}

\item{transformation}{choses type of transformation and back-transformation.}

\item{eff_smpsize}{Effective sample size.}

\item{MSE}{a character string determining the estimation method of the MSE.
Methods that can be chosen
(i) analytical MSE depending on the estimation method of the variance of the
random effect ("\code{analytical}"),
(ii) a jackknife MSE ("\code{jackknife}"),
(iii) and a bootstrap that returns the backtransformed confidence intervals
when the arcsin transformation is chosen ("\code{boot}").}

\item{B}{numeric value that determines the number of bootstrap iterations.}

\item{alpha}{a numeric value that determines the confidence level for the
confidence intervals.}
}
\value{
fitted FH model.
}
\description{
This function conducts the estimation of a Fay-Herriot model.
}
