% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH.R
\name{FH_AK}
\alias{FH_AK}
\title{Function for Fay-Herriot model}
\usage{
FH_AK(formula, vardir, combined_data, domains = NULL, method,
  back_transformation, interval = c(0, 1000), precision = 1e-04,
  maxiter = 100)
}
\arguments{
\item{formula}{formula object.}

\item{vardir}{direct variance.}

\item{combined_data}{combined data.}

\item{domains}{domain level.}

\item{method}{method for the estimation of sigmau2.}

\item{interval}{interval for the estimation of sigmau2.}

\item{precision}{precision criteria for the estimation of sigmau2.}

\item{maxiter}{maximum of iterations for the estimation of sigmau2.}
}
\value{
fitted FH model.
}
\description{
This function conducts the estimation of a Fay-Herriot model.
}
