% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FH_new.R
\name{FH_eblup}
\alias{FH_eblup}
\title{Function for Fay-Herriot model}
\usage{
FH_eblup(formula, vardir, combined_data, domains = NULL, method,
  transformation = "no", interval = c(0, 1000), precision = 1e-04,
  maxiter = 100)
}
\arguments{
\item{formula}{formula object.}

\item{vardir}{direct variance.}

\item{combined_data}{combined data.}

\item{domains}{domain level.}

\item{method}{method for the estimation of sigmau2.}

\item{transformation}{choses type of transformation and back-transformation.}

\item{interval}{interval for the estimation of sigmau2.}

\item{precision}{precision criteria for the estimation of sigmau2.}

\item{maxiter}{maximum of iterations for the estimation of sigmau2.}
}
\value{
fitted FH model.
}
\description{
This function conducts the estimation of a Fay-Herriot model.
}
