% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fh.R
\name{print.fh}
\alias{print.fh}
\title{Prints an emdiObject}
\usage{
\method{print}{fh}(x, ...)
}
\arguments{
\item{x}{an x of type "emdi", representing point and MSE
estimates obtained by direct estimation (see also \code{\link{direct}})
or Empirical Best Prediction (see also \code{\link{ebp}}).}

\item{...}{optional arguments passed to \code{\link{print.default}}.}
}
\description{
Basic information of an emdi object is printed.
}
\seealso{
\code{\link{emdiObject}}, \code{\link{ebp}}
}
