% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fh.R
\name{print.summary.fh}
\alias{print.summary.fh}
\title{Prints a summary.emdi object}
\usage{
\method{print}{summary.fh}(x)
}
\arguments{
\item{x}{an object of type "summary.emdi", generally resulting
from applying summary to an object of type "emdi"}

\item{...}{optional arguments passed to print.default; see the documentation on
that method functions.}
}
\description{
The elements described in summary.emdi are printed.
}
