% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_sigmau2.R
\name{saeReml}
\alias{saeReml}
\title{Function for estimating sigmau2 following the \pkg{sae} package.}
\usage{
saeReml(vardir, precision, maxiter, X, y)
}
\arguments{
\item{vardir}{direct variance.}

\item{precision}{precision criteria for the estimation of sigmau2.}

\item{maxiter}{maximum of iterations for the estimation of sigmau2.}

\item{X}{matrix with explanatory variables.}

\item{y}{direct estimator.}
}
\value{
estimated sigmau2.
}
\description{
This function estimates sigmau2.
}
\keyword{internal}
