% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fh.R
\name{summary.fh}
\alias{summary.fh}
\title{Summarizes an emdiObject}
\usage{
\method{summary}{fh}(object)
}
\arguments{
\item{object}{an object of type "emdi", representing point and MSE
estimates. Objects differ depending on the estimation method: direct
vs. model-based.}

\item{...}{additional arguments that are not used in this method.}
}
\value{
an object of type "summary.emdi" with following
}
\description{
Additional information about the data and model in small area estimation
methods and components of an emdi object are extracted. The returned object
is suitable for printing  with the \code{print.summary.emdi} method.
}
