% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh.R
\name{fh}
\alias{fh}
\title{Function for Fay-Herriot model}
\usage{
fh(fixed, vardir, combined_data, domains = NULL, method = "reml",
  transformation = "no", eff_smpsize = NULL, interval = c(0, 1000),
  precision = 1e-04, maxiter = 100, alpha = alpha)
}
\arguments{
\item{fixed}{a two-sided linear formula object describing the
fixed-effects part of the nested error linear regression model with the
dependent variable on the left of a ~ operator and the explanatory
variables on the right, separated by + operators.}

\item{vardir}{a character string indicating the name of the variable containing
the domain-specific sampling variances of the direct estimators that are
included in \cr \code{combined_data}.}

\item{combined_data}{a data set containing the direct estimates,
the sampling variances, the explanatory variables and the domains.}

\item{domains}{a character string indicating the domain variable that is
included in \cr \code{combined_data}. If \code{NULL}, the domains are numbered
consecutively.}

\item{method}{a character string describing the method for the estimation of
the variance of the random effects. Methods that can be chosen
(i) restricted maximum likelihood (REML) method ("\code{reml}"),
(iii) maximum likelihood method ("\code{ml}"),
(iv) adjusted REML following \cite{Li and Lahiri (2010)} ("\code{amrl}"),
(v) adjusted ML following \cite{Li and Lahiri (2010)} ("\code{ampl}"),
(vi) adjusted REML following \cite{Yoshimori and Lahiri (2014)}
("\code{amrl_yl}"),
(vii) adjusted ML following \cite{Yoshimori and Lahiri (2014)}
("\code{ampl_yl}").
 Defaults to "\code{reml}".}

\item{transformation}{choses type of transformation and back-transformation.}

\item{eff_smpsize}{Effective sample size.}

\item{interval}{interval for the estimation of sigmau2.}

\item{precision}{precision criteria for the estimation of sigmau2.}

\item{maxiter}{maximum of iterations for the estimation of sigmau2.}
}
\value{
fitted FH model.
}
\description{
This function conducts the estimation of a Fay-Herriot model.
}
